Re-init frontend for the 10th billion time while I try to learn CSS

+++ backend/src/main/java/com/plainoldmoose/IDLWebApp/dto/response/player/EloHistoryResponse.java
@@ -5,7 +5,7 @@ import com.plainoldmoose.IDLWebApp.model.enums.EloChangeReason;
 import java.time.LocalDateTime;
 
 public record EloHistoryResponse(
-        Long elo,
+        int elo,
         int eloChange,
         LocalDateTime timestamp,
         EloChangeReason reason,
